refactor(trie): optimize code, remove unused code
