+++ framework/src/test/java/org/tron/core/actuator/UnfreezeBalanceActuatorTest.java
@@ -164,10 +164,11 @@ public class UnfreezeBalanceActuatorTest extends BaseTest {
       Assert.assertEquals(totalNetWeightBefore,
           totalNetWeightAfter + frozenBalance / 1000_000L);
 
-      DynamicPropertiesStateStore stateStore = new DynamicPropertiesStateStore();
-      stateStore.init(queryInstance);
-      Assert.assertEquals(totalNetWeightBefore,
-              stateStore.getTotalNetWeight() + frozenBalance / 1000_000L);
+      try (DynamicPropertiesStateStore stateStore = new DynamicPropertiesStateStore()) {
+        stateStore.init(queryInstance);
+        Assert.assertEquals(totalNetWeightBefore,
+                stateStore.getTotalNetWeight() + frozenBalance / 1000_000L);
+      }
 
     } catch (ContractValidateException e) {
       Assert.assertFalse(e instanceof ContractValidateException);
